#76 Small src refactoring
