+++ tigra_software/launch/drivers/uc_socket_server.launch
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
 <launch>
-    <node pkg="rosserial_server" type="socket_node" name="rosserial_server" output="screen">
+    <node pkg="rosserial_server" type="socket_node" name="rosserial_server" output="screen" ns="tigra">
         <param name="port" value="23456" />
     </node>
 </launch>
